+++ modules/iam-identity/src/main/java/com/ibm/cloud/platform_services/iam_identity/v1/IamIdentity.java
@@ -12,7 +12,7 @@
  */
 
 /*
- * IBM OpenAPI SDK Code Generator Version: 3.113.0-3f9df07a-20260317-160650
+ * IBM OpenAPI SDK Code Generator Version: 3.113.1-d76630af-20260320-135953
  */
 
 package com.ibm.cloud.platform_services.iam_identity.v1;
@@ -1905,7 +1905,9 @@ public class IamIdentity extends BaseService {
     }
     builder.header("Accept", "application/json");
     final JsonObject contentJson = new JsonObject();
-    contentJson.addProperty("value_string", updatePreferenceOnScopeAccountOptions.valueString());
+    if (updatePreferenceOnScopeAccountOptions.valueString() != null) {
+      contentJson.addProperty("value_string", updatePreferenceOnScopeAccountOptions.valueString());
+    }
     if (updatePreferenceOnScopeAccountOptions.valueListOfStrings() != null) {
       contentJson.add("value_list_of_strings", com.ibm.cloud.sdk.core.util.GsonSingleton.getGson().toJsonTree(updatePreferenceOnScopeAccountOptions.valueListOfStrings()));
     }