+++ modules/iam-identity/src/main/java/com/ibm/cloud/platform_services/iam_identity/v1/model/CreateLinkOptions.java
@@ -172,7 +172,7 @@ public class CreateLinkOptions extends GenericModel {
   /**
    * Gets the crType.
    *
-   * The compute resource type. Valid values are VSI, IKS_SA, ROKS_SA.
+   * The compute resource type. Valid values are VSI, PVS, BMS, IKS_SA, ROKS_SA, CE.
    *
    * @return the crType
    */