+++ modules/iam-identity/src/main/java/com/ibm/cloud/platform_services/iam_identity/v1/model/UpdateClaimRuleOptions.java
@@ -356,7 +356,7 @@ public class UpdateClaimRuleOptions extends GenericModel {
    * Gets the crType.
    *
    * The compute resource type the rule applies to, required only if type is specified as 'Profile-CR'. Valid values are
-   * VSI, IKS_SA, ROKS_SA.
+   * VSI, PVS, BMS, IKS_SA, ROKS_SA, CE.
    *
    * @return the crType
    */